+++ tubes/view/admin-view-matkul.php
@@ -128,7 +128,7 @@ if($_SESSION['level']!="admin")
         <td>$row[nama_matkul]</td>
         <td>
         <a class='btn btn-success' href='../admin_function/admin-update-matkul.php?kode_matkul=$row[kode_matkul]'>Update</a>
-                  <a class='btn btn-danger' href='../admin_function/admin-delete.php?kode_matkul=$row[kode_matkul]'>Delete</a>
+                  <a class='btn btn-danger' href='../admin_function/admin-delete.php?kode_matkul=$row[kode_matkul]'onClick=\"return confirm('Do you want to Delete this data ?');\"'>Delete</a>
                 </td>
       </tr>
       ";
